+++ CHANGELOG.md
@@ -15,7 +15,20 @@ The project uses semantic versioning (see [semver](https://semver.org)).
   - New TOML file entry `[mass-driver.scan]`, with subkey `scanner_names` used
     to select which scanner plugins to enable.
   - Simple scanners `root-files` and `dockerfile-from` provided for reference
-  - New fixture `massdriver_scan` for testing scanners
+  - New fixture `massdriver_scan` and `massdriver_scan_check` for testing scanners
+
+### Changed
+- Test fixture `massdrive_check_file` now returns unchecked `result` and
+  `reference` blobs for equality assertion (`assert result == reference`) to be
+  done by the caller. This enables plugins like `pytest-clarity` to show
+  colorful diff. Users of `massdrive_check_file` need to change (on pain of lack
+  of test assertion):
+
+```diff
+- massdrive_check_file(workdir)
++ result, reference = massdrive_check_file(workdir)
++ assert result == reference, "Massdriver result should match reference"
+```
 
 
 ## v0.12.0 - 2023-04-05